backend_2fa_admin/requirements.txt
aio-pika==9.3.1
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==3.2.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
'faststream[rabbit]'
frozenlist==1.5.0
h11==0.14.0
httpx==0.27.0
idna==3.10
magic-filter==1.0.12
motor==3.7.0
multidict==6.3.0
passlib==1.7.4
propcache==0.3.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT>=2.10.1
pymongo==4.11.3
python-dotenv==0.21.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2025.2
requests==2.31.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.13.0
uvicorn==0.19.0
yarl==1.18.3 